pollpulse-tn-main/pollpulse-tn-main/pollpulse-tn-main/requirements.txt
# YouTube scraping - using yt-dlp
yt-dlp>=2024.1.0
youtube-transcript-api>=0.6.0

# Supabase integration
supabase>=2.0.0
python-dotenv>=1.0.0

# Machine Learning
transformers>=4.30.0
torch>=2.0.0
protobuf>=3.20.0
sentencepiece>=0.2.1
tokenizers>=0.13.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Data processing
pandas>=2.3.3
numpy>=2.4.0

# Date parsing
dateparser>=1.2.2

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
